doc: Extend Claude instructions
Extend development process by STR creation step. Having the test matrix in the STP turned out to be a hassle: At the moment when we want to release the document, we have not yet written any tests. If the set of implemented tests does not match the predicted set, the STP has to be changed. And the test result can only be added after having released the STP. Or the release of the STP would need to be delayed.

Sharpen the first coding activity in the development process. "Some construction" makes sense for humans to get a rough understanding of the domain. At cost of a few iterations in the Construction <=> Testing loop, an ML model does not need to do "some construction" upfront. Instead, the new step aims to help the human user to build up a rough understanding despite not writing code.

Demand claude to re-scan artifacts under review upon feedback or approval